Move loop compilation to general utils, rename bindings
