+++ test/include/test_utils.h
@@ -19,31 +19,9 @@
 #include "reffine/pass/printer.h"
 #include "reffine/pass/reffinepass.h"
 #include "reffine/pass/scalarpass.h"
+#include "reffine/utils/utils.h"
 
 arrow::Result<reffine::ArrowTable> get_input_vector();
 std::string print_arrow_table(reffine::ArrowTable&);
-
-template <typename T>
-T compile_loop(std::shared_ptr<reffine::Func> loop)
-{
-    reffine::CanonPass::Build(loop);
-    auto exp_loop = reffine::LoadStoreExpand::Build(loop);
-    auto elm_loop = reffine::NewGetElimination::Build(exp_loop);
-
-    auto jit = reffine::ExecEngine::Get();
-    auto llmod = make_unique<llvm::Module>("test", jit->GetCtx());
-
-    reffine::LLVMGen::Build(elm_loop, *llmod);
-    jit->AddModule(std::move(llmod));
-    return jit->Lookup<T>(loop->name);
-}
-
-template <typename T>
-T compile_op(std::shared_ptr<reffine::Func> op)
-{
-    auto op_to_loop = reffine::OpToLoop::Build(op);
-    auto loop = reffine::LoopGen::Build(op_to_loop);
-    return compile_loop<T>(loop);
-}
 
 #endif  // TEST_INCLUDE_TEST_UTILS_H_
